Add permissions for webapp user
We can restrict permissions to 775 and change ownership to webapp. When deployment is happening and there is some log entries in laravel.log, that file is owned by root by default. That means when actual application log is coming from laravel (which is under webapp user), permission to access it is denied if it's 775.

This solution should restrict permission reasonably with laravel still able to write.